+++ app/src/main/java/com/r0adkll/deckbuilder/arch/domain/features/collection/model/CollectionCount.kt
@@ -9,5 +9,6 @@ data class CollectionCount(
         val id: String,
         val count: Int,
         val set: String,
-        val series: String
+        val series: String,
+        val isSourceOld: Boolean = false
 ) : Parcelable